pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crop-disease-classification"
version = "1.0.0"
description = "A machine learning project that classifies healthy vs unhealthy wheat and maize crops using CNN"
authors = [
    {name = "Vijay", email = "vijay@example.com"},
]
dependencies = [
    "tensorflow>=2.13.0",
    "opencv-python>=4.8.0",
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.3.0",
    "kaggle>=0.1.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]
requires-python = ">=3.8.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/vijay/crop-disease-classification"
Repository = "https://github.com/vijay/crop-disease-classification.git"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0,<7.2.0",
    "jupyter>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
